+++ htdocs/components/ActivityComponent/subComponents/ActivityEditComponent.js
@@ -86,7 +86,8 @@ class ActivityEditComponent extends MainComponent {
                 customerNotesTemplate: "",
                 priorityChangeReason: "",
                 emptyAssetReason: "",
-                emptyAssetReasonNotify: false
+                emptyAssetReasonNotify: false,
+                Inbound:null
             },
             currentActivity: "",
             _showModal: false,
@@ -243,7 +244,7 @@ class ActivityEditComponent extends MainComponent {
         delete data.activities;
         delete data.onSiteActivities;
         delete data.documents;
-        const finalData = pick(data, [
+        const finalData = pick(data, [            
             "callActivityID",
             "alarmDate",
             "alarmTime",
@@ -271,9 +272,10 @@ class ActivityEditComponent extends MainComponent {
             "hideFromCustomerFlag",
             "submitAsOvertime",
             "emptyAssetReason",
-            "completeDate"
+            "completeDate",
+            "Inbound"
         ]);
-
+        console.log(finalData);
         this.api
             .updateActivity(finalData)
             .then((response) => {
@@ -1088,27 +1090,85 @@ class ActivityEditComponent extends MainComponent {
         }
 
         return this.getElementControl(
-            "Type",
-            "Type",
+          "Type",
+          "Type",
+          el(
+            "div",{style:{display:"flex",flexDirection:"row"}},
             el(
-                "select",
-                {
-                    disabled: !isEnabled,
-                    required: true,
-                    value: data?.callActTypeID || "",
-                    onChange: (event) =>
-                        this.setValue("callActTypeID", event.target.value),
-                    style: {width: "100%"}
-                },
-                el("option", {key: "empty", value: ""}, "Please select"),
-                activityTypesToShow.map((t) =>
-                    el("option", {key: t.id, value: t.id}, t.description)
-                )
-            )
+              "select",
+              {
+                disabled: !isEnabled,
+                required: true,
+                value: data?.callActTypeID || "",
+                onChange: (event) => this.handleTypeChange(event.target.value),
+                style: { width: "100%" },
+              },
+              el("option", { key: "empty", value: "" }, "Please select"),
+              activityTypesToShow.map((t) =>
+                el("option", { key: t.id, value: t.id }, t.description)
+              )
+            ),
+            this.getInboundIcon()
+          )
         );
     };
 
-
+    handleTypeChange=(value)=>{
+        this.setValue("callActTypeID", value);
+        if(value=='11')
+        this.setState({showInboundOutboundModal:true});
+        else
+        this.setValue("Inbound", null);
+    }
+    getInboundOutBoundModal=()=>{
+        const {data}=this.state;
+        const Inbound=data.Inbound==null?false:data.Inbound;
+        const Outbound=data.Inbound==null?false:!data.Inbound;
+        return <Modal 
+        width={300}
+        show={this.state.showInboundOutboundModal}
+        title="Select contact type"
+        footer={<div key="footerActions" >            
+            <button  onClick={()=>this.setState({showInboundOutboundModal:false})}>OK</button>
+            <button  onClick={()=>this.setState({showInboundOutboundModal:false,Inbound:null})}>Cancel</button>
+        </div>}
+        >
+            <div style={{display:'flex', flexDirection:'row',justifyContent:"space-between"}}>
+                <div>
+                    <label className="mr-2">Inbound</label>
+                    <Toggle checked={Inbound} onChange={()=>this.setValue("Inbound",true)}></Toggle>
+                </div>
+                <div>
+                    <label className="mr-2">Outbound</label>
+                    <Toggle  checked={Outbound}  onChange={()=>this.setValue("Inbound",false)}></Toggle>
+                </div>
+            </div>
+        </Modal>
+    }
+    getInboundIcon=()=>{
+        const { data } = this.state;
+        switch (data.Inbound) {
+          case true:
+            return (
+              <ToolTip title="Inbound Contact" width={15}>
+                  <i  onClick={() =>
+                    this.setState({ showInboundOutboundModal: true })
+                  } className="fal fa-sign-in pointer icon"></i>                 
+              </ToolTip>
+            );
+          case false:
+            return (
+              <ToolTip title="Outbound Contact" width={15}>
+                  <i onClick={() =>
+                    this.setState({ showInboundOutboundModal: true })
+                  } className="fal fa-sign-out  pointer icon"></i>
+                
+              </ToolTip>
+            );
+          default:
+            return null;
+        }
+    }
     getContactsElement = () => {
         const {el} = this;
         const {data, contacts, currentContact} = this.state;
@@ -1938,12 +1998,13 @@ class ActivityEditComponent extends MainComponent {
         return (
             <div style={{width: "90%"}}>
                 {this.getAdditionalChargeModal()}
+                {this.getInboundOutBoundModal()}
                 {this.getAlert()}
                 {this.getConfirm()}
                 {this.getPrompt()}
                 {this.getPriorityChangeReason()}
                 {this.getProjectsElement()}
-                {this.getCallbackModal()}
+                {this.getCallbackModal()}                                
                 <ActivityHeaderComponent serviceRequestData={data}/>
                 <div className="activities-edit-container">
                     {this.getActions()}